requirements.txt
selenium
requests
mysql-connector-python
python-dotenv